+++ helper/concurrentmap/concurrent_map.go
@@ -1,4 +1,4 @@
-package common
+package concurrentmap
 
 import "sync"
