+++ helper/concurrentmap/concurrent_map_test.go
@@ -1,4 +1,4 @@
-package common
+package concurrentmap
 
 import "testing"
