+++ protocol/syncer.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/dogechain-lab/dogechain/blockchain"
-	"github.com/dogechain-lab/dogechain/helper/common"
+	cmap "github.com/dogechain-lab/dogechain/helper/concurrentmap"
 	"github.com/dogechain-lab/dogechain/helper/progress"
 	"github.com/dogechain-lab/dogechain/network"
 	"github.com/dogechain-lab/dogechain/network/event"
@@ -65,7 +65,7 @@ type Syncer struct {
 	logger     hclog.Logger
 	blockchain blockchainShim
 
-	peers common.ConcurrentMap // Maps peer.ID -> SyncPeer
+	peers cmap.ConcurrentMap // Maps peer.ID -> SyncPeer
 
 	serviceV1 *serviceV1
 	stopCh    chan struct{}
@@ -86,7 +86,7 @@ func NewSyncer(logger hclog.Logger, server *network.Server, blockchain blockchai
 		blockchain:      blockchain,
 		server:          server,
 		syncProgression: progress.NewProgressionWrapper(progress.ChainSyncBulk),
-		peers:           common.NewConcurrentMap(),
+		peers:           cmap.NewConcurrentMap(),
 	}
 
 	return s
